refactor(tui): extract mapping and actions from app.rs
Split 781-line app.rs into focused modules:
- mapping.rs: event-to-Msg translation (305 lines)
- actions.rs: send_message, tab completion, scroll helpers (149 lines)
- app.rs: struct + lifecycle + dispatch (332 lines)

Uses Rust's multiple impl blocks — methods stay on App,
just defined in separate files. No calling code changes.

Pure refactor — zero behavior changes.
